+++ Sources/BottomSheet/BottomSheetView/BottomSheetView+Gestures.swift
@@ -14,8 +14,8 @@ internal extension BottomSheetView {
                 // Perform custom onChanged action
                 self.configuration.onDragChanged(value)
                 
-                // Update translation; on iPad and Mac the drag direction is reversed
-                self.translation = self.isIPadOrMac ? -value.translation.height : value.translation.height
+                // Update translation; on iPad floating and Mac the drag direction is reversed
+                self.translation = self.isIPadFloatingOrMac ? -value.translation.height : value.translation.height
                 // Dismiss the keyboard on drag
                 self.endEditing()
             }
@@ -51,8 +51,8 @@ internal extension BottomSheetView {
                     
                     // Notify the ScrollView that the user is dragging
                     self.dragState = .none
-                    // Update translation; on iPad and Mac the drag direction is reversed
-                    self.translation = self.isIPadOrMac ? -value.translation.height : value.translation.height
+                    // Update translation; on iPad floating and Mac the drag direction is reversed
+                    self.translation = self.isIPadFloatingOrMac ? -value.translation.height : value.translation.height
                 }
                 
                 // Dismiss the keyboard on dragging/scrolling